Massive rearrangement of code
